Feature/task 1 user entity

+++ src/main/java/com/example/taskmanagementapi/service/TaskServiceImpl.java
@@ -3,7 +3,8 @@ package com.example.taskmanagementapi.service;
 import com.example.taskmanagementapi.dto.TaskDto;
 import com.example.taskmanagementapi.entity.Task;
 
-
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import com.example.taskmanagementapi.exception.ResourceNotFoundException;
 import com.example.taskmanagementapi.mapper.TaskMapper;
 import com.example.taskmanagementapi.repository.TaskRepository;
@@ -16,6 +17,8 @@ public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
+    private static final Logger log = LoggerFactory.getLogger(TaskServiceImpl.class);
+
 
     public TaskServiceImpl(TaskRepository taskRepository, TaskMapper taskMapper){
         this.taskRepository = taskRepository;
@@ -24,40 +27,65 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Page<TaskDto> getAllTasks(Pageable pageable) {
+        log.info("Fetching all tasks");
         return taskRepository.findAll(pageable)
                 .map(taskMapper::toDto);
     }
 
     @Override
     public TaskDto getTaskById(Long id) {
+        log.info("Fetching task with id {}", id);
+
         Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Task not found"));
+                .orElseThrow(() -> {
+                    log.warn("Task not found with id {}", id);
+                    return new ResourceNotFoundException("Task not found");
+                });
         return taskMapper.toDto(task);
     }
 
     @Override
     public TaskDto createTask(TaskDto taskDto){
+        log.info("Creating task with title {}", taskDto.getTitle());
+
         Task task = taskMapper.toEntity(taskDto);
         Task savedTask = taskRepository.save(task);
+
+        log.info("Task created with id {}", savedTask.getId());
+
         return taskMapper.toDto(savedTask);
     }
 
     @Override
     public TaskDto updateTask(Long id, TaskDto taskDto) {
+        log.info("Updating task with id {}", id);
+
         Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Task not found"));
+                .orElseThrow(() ->  {
+                    log.warn("Task not found for updated with id {}",id);
+                    return new ResourceNotFoundException("Task not found");
+                });
 
         taskMapper.updateEntity(task, taskDto);
 
         Task updatedTask = taskRepository.save(task);
+
+        log.info("Task updated with id {}", id);
 
         return taskMapper.toDto(updatedTask);
     }
 
     @Override
     public void deleteTask(Long id) {
+        log.info("Deleting task with id {}", id);
+
+        if (!taskRepository.existsById(id)) {
+            log.warn("Task not found for deletion with id {}", id);
+            throw new ResourceNotFoundException("Task not found");
+        }
+
         taskRepository.deleteById(id);
+        log.info("Task deleted with id {}", id);
     }
 
-    String status = "CLOSED";
 }
